subprocess.run or subprocess.call?
Just wanted to verify that we should be using call or run? Earlier in a
comment on my PR you indiciated run but I saw in your latest copy that
it is now call. Just wanted to make sure this wasn't an unintentional
ommission.

Also, I saw you implemented a string split operation instead of
os.path.basename to get the basename of the file to feed ffmpeg. Even
though I agree that string split probably should do the job, I prefer we
stick to the os.path functions if only because then we can leverage
their (the python devs) work to ensure we always get the correct basename accross
platforms and etc. So we can just not have to worry if it works. I also changed the
import os to from os.path import split to only import what we need.