Filter folder contents before iterating for inference
Instead of two if statements to be run for each inference loop, I
changed it to first delete the path we know must exist, which is the
master audio file, and thus don't really have to check if it exists. And
then it quickly iterates through all the  remaining paths and removes
any paths that don't start with the prefix before running the main
inference loop which now only contains paths that are relevant.

Both of these are not strictly neccessary, but I think it looks cleaner
and more efficiently segments code. If in the future we want to run more
operations on the audio segments now we only have one line in that loop
instead of all this other stuff on top of it. Each extra nested level is
another tab we have to keep visual track of and hence my preference for
flatter nests when possible.